+++ src/main/java/com/oneandone/sushi/metadata/simpletypes/FloatType.java
@@ -33,7 +33,7 @@ public class FloatType extends SimpleType {
 
     @Override
     public String valueToString(Object obj) {
-        return ((Float) obj).toString();
+        return obj.toString();
     }
     
     @Override
